Student Class, into list: shuffle, sort, and Info
The code above is capable of making a class of student and creating their instances, they are able to be given their own attributes such as name, age, and id_number. 
As per instruction the program is capable of adding more objects(the students) and putting them into a list that can now be shuffled randomly and sorted through age, name, and id number. Lastly the program is capable of printing or displaying the students and their respective information.